requirements.txt
gymnasium[atari,accept-rom-license]==0.29.1
h5py==3.11.0
huggingface-hub==0.17.2
hydra-core==1.3
numpy==1.26.0
opencv-python==4.10.0.84
pillow==10.3.0
pygame==2.5.2
# torch==2.4.1
torcheval==0.0.7
tqdm==4.66.4
wandb==0.17.0
einops